# util: expose promise events