--- a/lib/util.js
+++ b/lib/util.js
@@ -21,6 +21,7 @@
 
 'use strict';
 
+const { internalBinding } = require('internal/bootstrap/loaders');
 const errors = require('internal/errors');
 const {
   ERR_FALSY_VALUE_REJECTION,
@@ -37,8 +38,8 @@ const {
   kRejected,
   previewEntries
 } = process.binding('util');
+const { setPromiseHook } = internalBinding('safe_util');
 
-const { internalBinding } = require('internal/bootstrap/loaders');
 const types = internalBinding('types');
 Object.assign(types, require('internal/util/types'));
 const {
@@ -1464,6 +1465,72 @@ function getSystemErrorName(err) {
   return internalErrorName(err);
 }
 
+const promiseEventTypes = [
+  'init',
+  'resolve',
+  'before',
+  'after',
+  'rejectWithNoHandler',
+  'handlerAddedAfterReject',
+  'rejectAfterResolved',
+  'resolveAfterResolved',
+];
+const promiseHooks = new Set();
+let promiseHookWarned = false;
+let resolveHookRef = 0;
+function promiseHookCallback(type, promise, value) {
+  type = promiseEventTypes[type];
+  for (const hooks of promiseHooks) {
+    try {
+      const hook = hooks[type];
+      if (typeof hook === 'function') {
+        hook(promise, value);
+      }
+    } catch (e) {
+      process.nextTick(() => { throw e; });
+    }
+  }
+}
+function createPromiseHook(hooks) {
+  if (!promiseHookWarned) {
+    promiseHookWarned = true;
+    process.emitWarning('The util.createPromiseHook API is experimental.',
+                        'ExperimentalWarning');
+  }
+
+  const hasResolve = 'resolve' in hooks;
+
+  promiseEventTypes.forEach((type) => {
+    const hook = hooks[type];
+    if (hook === undefined) {
+      return;
+    }
+
+    if (typeof hook !== 'function') {
+      throw new ERR_INVALID_ARG_TYPE(`hooks.${type}`, 'function', hook);
+    }
+  });
+
+  return {
+    enable() {
+      promiseHooks.add(hooks);
+      if (hasResolve) {
+        resolveHookRef++;
+      }
+      setPromiseHook(promiseHookCallback, resolveHookRef > 0);
+    },
+    disable() {
+      promiseHooks.delete(hooks);
+      if (hasResolve) {
+        resolveHookRef--;
+      }
+      setPromiseHook(
+        hooks.size === 0 ? null : promiseHookCallback,
+        resolveHookRef > 0);
+    },
+  };
+}
+
 // Keep the `exports =` so that various functions can still be monkeypatched
 module.exports = exports = {
   _errnoException: errors.errnoException,
@@ -1504,6 +1571,8 @@ module.exports = exports = {
   TextDecoder,
   TextEncoder,
   types,
+
+  createPromiseHook,
 
   // Deprecated Old Stuff
   debug: deprecate(debug,