@benjamingr 2018-07-17
> This allows the default behaviour in node to follow the intention of js

This is also not true and was not the intention - browsers error on unhandled rejections too. One of the things we've been getting since adding the `unhandledRejection` hook is constant positive feedback for logging as well as negative one for not crashing.