@benjamingr 2018-07-17
By hte way: By the same logic we should deprecate `uncaughtException`, stop exiting or logging on errors and expose an `async_hook` for exiting.