@benjamingr 2018-07-17
> they are explicitly intended to be able to silently fail.

I do not believe this is the case nor does the spec make any such requirement.  That said: Under all proposed implementations promises are able to silently fail - just like you can ignore synchronous errors:

```js
process.on('uncaughtException', e => { /* console.log('https://i.imgur.com/c4jt321.png') */ });
```

You can ignore unhandled rejections:

```js
process.on('uncaughtException', e => { /* console.log('https://i.imgur.com/c4jt321.png') */ });
```

Which will continue to work forever under all proposed changes. The question is about what the _default_ behaviour should be.

We have a consensus at this point in Node.js that ignoring uncaught errors should not be the default behaviour. 

> what browser consoles do is completely irrelevant because that's a debug facility. this is all for debugging

Again, the same argument could be made for Node.js not exiting on uncaught errors. Since a significant part of our user-base use `async/await` it means that in practice they don't distinguish between unhandled rejections and uncaught errors in terms of how they handle errors anyway.