@devsnek 2018-07-17
> I do not believe this is the case nor does the spec make any such requirement.

the spec says `An implementation of HostPromiseRejectionTracker must complete normally in all cases.`. i would argue that killing the process is not a normal completion, but i don't really want to get into these arguments.

aside from that, `unhandledRejection` can't guarantee that it will be fired for the things it should be unless someone is using `global.gc`, which kind of undermines the debugging potential of it. By not making these (arguably confusing for that reason) apis, we can let people find the best behaviour for what they're trying to do, similar to how people often use tools like winston even though node has `console`