@devsnek 2018-07-17
> Also note this cannot be implemented in user-land since it relies on core GC behaviour.

the destroy hook lets us do that right?