@benjamingr 2018-07-22
@AndreasMadsen well, promises are different from other stuff `async_hooks` are used for anyway - even the existing `resolve` hook. This is a little footgunny - a recent example is jest listening to all async hooks and giving an error because there was a pending unresolved promise (which is totally fine unlike an open file handle for example).