@mcollina 2018-07-23
I'm tagging this semver-major because async_hooks is used in the wild quite a bit, and I would prefer this not to get into existing lines.

Apart from this, I'm -0 with the changes.