@devsnek 2018-07-23
my current thought on what to do with this is `require('util').addPromiseHook(callback)` (and removePromiseHook).

does anyone have a more idiomatic idea?