@benjamingr 2018-07-23
> does anyone have a more idiomatic idea?

Going to try another ping to @nodejs/async_hooks 

I think additions of a method to process would require TSC approval and personally I think this is as much `async_hooks` as the resolve hook. 

This has no -1s and the code changes LGTM (both under async hooks and separately), I wanted Ruben's feedback on it though too.