@AndreasMadsen 2018-07-23
Okay, I will give it a -1 just to be clear about my opinion.

If you want to change my opinion, then please explain it's relation to async resources in general. Without defeing to argument such as "well we already have promiseResolve".