@mcollina 2018-08-09
I do not know why, but landing this commit introduces a 10-20% throughput drop even if the promise hooks are disabled. Adding this feature should not have _any_ overhead if disabled.
This is different from the issue I've flagged earlier, which is resolved. Listening to only `'rejectAfterResolved'` and `'resolvedAfterResolved'` does not introduce _any_ more overhead.

I've used the following server to verify:

```js
'use strict'

const Hapi = require('hapi')

// require('util').createPromiseHook((type, promise, value) => {
// }, ['rejectAfterResolved', 'resolveAfterResolved']).enable()

async function start () {
  const server = Hapi.server({ port: 3000, debug: false })

  server.route({
    method: 'GET',
    path: '/',
    config: {
      cache: false,
      response: {
        ranges: false
      },
      state: { parse: false }
    },
    handler: async function (request, h) {
      return { hello: 'world' }
    }
  })

  await server.start()
}

start()
```

On master (2nd run to allow turbofan warm up):

```
$ autocannon -c 100 -d 5 localhost:3000
Running 5s test @ http://localhost:3000
100 connections

Stat         Avg     Stdev   Max
Latency (ms) 8.68    2.92    106.58
Req/Sec      10901.6 381.95  11286
Bytes/Sec    1.79 MB 76.4 kB 1.85 MB

55k requests in 5s, 8.94 MB read
```

On this branch (2nd run):

```
$ autocannon -c 100 -d 5 localhost:3000
Running 5s test @ http://localhost:3000
100 connections

Stat         Avg     Stdev   Max
Latency (ms) 10.68   3.59    122.03
Req/Sec      8924    369.32  9235
Bytes/Sec    1.46 MB 76.4 kB 1.51 MB

45k requests in 5s, 7.32 MB read
```

I do not know if we have any async/await heavy micro benchmarks.